refactor(Agendamento)
